pyproject.toml
[tool.poetry]
name = "progame"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.2.4"
certifi = "2019.11.28"
chardet = "3.0.4"
defusedxml = "0.6.0"
Django = "3.0.4"
django-allauth = "0.41.0"
django-bootstrap-modal-forms = "1.5.0"
django-ckeditor = "5.9.0"
django-crispy-forms = "1.9.0"
django-extra-views = "0.13.0"
django-filter = "2.2.0"
django-js-asset = "1.2.2"
djangorestframework = "3.11.0"
gunicorn = "20.0.4"
idna = "2.9"
mysqlclient = "1.4.6"
oauthlib = "3.1.0"
Pillow = "7.0.0"
python3-openid = "3.1.0"
pytz = "2019.3"
requests = "2.23.0"
requests-oauthlib = "1.3.0"
six = "1.14.0"
sqlparse = "0.3.1"
urllib3 = "1.25.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
